fix login add QRcode fix admin-service issues

+++ source/admin-service/src/main/java/org/vision/service/admin/persistence/model/VisionClient.java
@@ -14,7 +14,7 @@ public class VisionClient implements Serializable {
 
     private String name;
 
-    private Byte gender;
+    private Integer gender;
 
     private Integer age;
